pyproject.toml
[tool.poetry]
name = "prometheus-query-builder"
version = "0.2.0"
description = "Package for generation Prometheus Query Language queries."
authors = ["Michail Tsyganov <tsyganov.michail@yandex.ru>"]
homepage = "https://github.com/m-chrome/prometheus-query-builder"
repository = "https://github.com/m-chrome/prometheus-query-builder"
readme = "README.md"
license = "MIT"
keywords = ["prometheus", "promql", "query-builder"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
pytest-cov = "^3.0.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/m-chrome/prometheus-query-builder/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
